# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.ErrorResponse:
    properties:
      error:
        example: 'failed to read uploaded image: ...'
        type: string
    type: object
...
  handler.LoginResponse:
    properties:
      expires_at:
        description: 过期时间（access 或 mfa_token 的）
        type: string
      mfa_required:
        description: MFARequired 为 true 时，需带上 mfa_token 调用 POST /auth/mfa/verify 完成登录。
        type: boolean
      mfa_token:
        description: MFAToken 是第二步验证用的短期令牌（5 分钟，一次性）。
        type: string
      token:
        description: 'JWT（放在 Authorization: Bearer <token> 头中）'
        type: string
      token_type:
        description: 恒为 Bearer
        type: string
      user:
        allOf:
        - $ref: '#/definitions/user.SafeUser'
        description: 用户信息（需要 MFA 第二步时不返回）
    type: object
  handler.MFADisableRequest:
    properties:
      code:
        description: 当前验证码
        example: "123456"
        type: string
      password:
        description: 当前密码
        example: secret123
        type: string
    required:
    - code
    - password
    type: object
  handler.MFAEnableRequest:
    properties:
      code:
        description: 认证器 App 当前 6 位验证码
        example: "123456"
        type: string
      secret:
        description: setup 返回的 secret
        example: JBSWY3DPEHPK3PXP
        type: string
    required:
    - code
    - secret
    type: object
  handler.MFAEnableResponse:
    properties:
      message:
        description: 提示信息
        type: string
    type: object
  handler.MFAResetConfirmRequest:
    properties:
      code:
        description: 邮箱收到的 6 位验证码
        example: "123456"
        type: string
      email:
        description: 账号绑定的邮箱
        example: user@example.com
        type: string
    required:
    - code
    - email
    type: object
  handler.MFAResetRequest:
    properties:
      email:
        description: 账号绑定的邮箱
        example: user@example.com
        type: string
      password:
        description: 账号密码
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  handler.MFASetupResponse:
    properties:
      otpauth_url:
        description: otpauth:// URI，可自行生成二维码
        type: string
      qr_code:
        description: data:image/png;base64,... 可直接放进 <img src>
        type: string
      secret:
        description: base32 密钥（手动输入时使用）
        type: string
    type: object
  handler.MFAVerifyRequest:
    properties:
      code:
        description: 认证器 App 当前 6 位验证码
        example: "123456"
        type: string
    required:
    - code
    type: object
...
  user.SafeUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      mfa_enabled:
        description: 是否已开启两步验证
        type: boolean
      nickname:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: wechatapp-web
    url: http://www.swagger.io/support
  description: |-
    Image background replacement API. Uploads an image, removes its
    background via the remove.bg service, then optionally composites
    the cutout onto a new background (solid color or another image).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: wechatapp-web API
  version: 1.0.0
paths:
...
  /auth/mfa:
    delete:
      consumes:
      - application/json
      description: |-
        Turns off two-factor authentication after verifying the
        account password and a current TOTP code.
      parameters:
      - description: Password and current code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.MFADisableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 两步验证已关闭'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Wrong password or code
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable MFA
      tags:
      - mfa
  /auth/mfa/enable:
    post:
      consumes:
      - application/json
      description: |-
        Verifies the submitted code against the setup secret and
        stores it, turning on two-factor authentication.
      parameters:
      - description: Secret from setup and a current code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.MFAEnableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: MFA enabled
          schema:
            $ref: '#/definitions/handler.MFAEnableResponse'
        "400":
          description: Invalid secret or code
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm and enable MFA
      tags:
      - mfa
  /auth/mfa/reset/confirm:
    post:
      consumes:
      - application/json
      description: |-
        Verifies the emailed code and disables two-factor
        authentication so the user can log in again with just their
        password and re-enroll a fresh authenticator.
      parameters:
      - description: Email and emailed code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.MFAResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 两步验证已重置'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Wrong/expired code or MFA not enabled
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Confirm an MFA reset with the emailed code
      tags:
      - mfa
  /auth/mfa/reset/request:
    post:
      consumes:
      - application/json
      description: |-
        When a user lost their authenticator, this sends a 6-digit
        verification code to the account email. Requires the account
        password. The response is generic to avoid account enumeration.
      parameters:
      - description: Email and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.MFAResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Generic success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Send an MFA reset code by email
      tags:
      - mfa
  /auth/mfa/setup:
    post:
      description: |-
        Generates a TOTP secret and provisioning URI/QR code. Nothing
        is persisted until /auth/mfa/enable confirms a valid code.
      produces:
      - application/json
      responses:
        "200":
          description: Secret, otpauth URL and QR code
          schema:
            $ref: '#/definitions/handler.MFASetupResponse'
        "400":
          description: MFA already enabled
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Begin MFA setup (生成 TOTP 密钥)
      tags:
      - mfa
  /auth/mfa/verify:
    post:
      consumes:
      - application/json
      description: |-
        Exchanges the short-lived mfa_token returned by /auth/login
        plus a TOTP code for a full access token.
      parameters:
      - description: TOTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.MFAVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Full access token
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Invalid code
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing/invalid/expired MFA challenge token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete login with the second factor
      tags:
      - mfa
...
  /users/{id}/mfa:
    delete:
      description: |-
        Disables two-factor authentication for another account, for
        example when the user lost their authenticator. Admin only.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: 已重置该用户的两步验证'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Not an admin
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reset a user's MFA (admin)
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT issued by POST /auth/login, sent as "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"